// src/types/TaxRate.tsx
{/*  ===================================THIS FILE WAS AUTO GENERATED=================================== */}

export type TaxRate = {
    taxID: number;
    code: string;
    name: string;
    countryCode: string;
    provinceCode: string;
    rate: number;
    effectiveFrom: string;
    effectiveTo: string;
    isActive: boolean;
    capturedDate: string;
    capturedBy: string;
};